workspace/MAYA/maya_b-main/maya_b-main/requirements.txt
fastapi[standard]
SQLAlchemy
psycopg2-binary
uvicorn[standard]
pydantic-settings
python-dotenv